+++ RELEASE_NOTES.md
@@ -1,6 +1,6 @@
 # Community Solid Server release notes
 
-## V4.0.1
+## v4.0.1
 Freezes the `oidc-provider` dependency to prevent a potential issue with the solid authn client
 as described in https://github.com/inrupt/solid-client-authn-js/issues/2103.